flask-app/requirements.txt
boto3
boto3[crt]
requests
pdfkit
Jinja2
python-dotenv
flask===1.1.2
gunicorn